[names-refactor] Deleted contributions landing reducers and actions.
