+++ assets/pages/contributions-landing/contributionsLandingReducers.js
@@ -8,7 +8,7 @@ import type { Contrib, ContribError, Amounts } from 'helpers/contributions';
 
 import { parse as parseContribution } from 'helpers/contributions';
 import { abTestReducer as abTests } from 'helpers/abtest';
-import type { Action } from '../actions/contributionsLandingActions';
+import type { Action } from './contributionsLandingActions';
 
 
 // ----- Types ----- //
